Fix a few spelling and grammar issues in the README.
Fixes a few grammar and spelling issues in the README.
For example, there is an awkward comma in the opening sentence.  Both "Packagist.org" and "packagist.org" were used.  One or the other should be used for consistency.

Use present tense in why to use a composer badge.  There was a run-on sentence in the "Contribution" section.

Other than that the changes are just the addition or removal of commas and the like.