Remove repo-access when not required
Currently only the badges for composer.lock and .gitattributes files
require actual access to the repository. That is currently only possible
for packages hosted on github or bitbucket.

Due to that requirement though all other badges will also not work for
packages hosted elsewhere.

This commit fixes that by decoupling the requirement for a default
branch from all the other badges that can be retrieved purely from
information available at packagist.